refactor: move the scan page guide chrome and the scene content types out
ScanPage hands its ten presentational pieces to ScanPageGuide (with a sibling
test) and keeps the state machine; SceneContent hands its prop and overlay
interfaces to sceneContentTypes and re-exports SceneContentProps.
